requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
psycopg2-binary
